Прошу проверить ДЗ. Версия 2.0

+++ src/main/java/homeWork2_6/employeeRecords/service/DepartmentService.java
@@ -6,13 +6,13 @@ import java.util.List;
 import java.util.Map;
 
 public interface DepartmentService {
-    double getMaxSalaryByDepartment(short departmentId);
+    double getMaxSalaryByDepartment(int departmentId);
 
-    double getMinSalaryByDepartment(short departmentId);
+    double getMinSalaryByDepartment(int departmentId);
 
-    List<Employee> getAllEmployeesByDepartment(short departmentId);
+    List<Employee> getAllEmployeesByDepartment(int departmentId);
 
-    Map<Short, List<Employee>> getAllByDepartment();
+    Map<Integer, List<Employee>> getAllByDepartment();
 
-    double getSumSalaryByDepartment(short departmentId);
+    double getSumSalaryByDepartment(int departmentId);
 }
